requirement.txt
absl-py==2.1.0
aggdraw==1.3.19
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attr==0.3.1
attrs==23.2.0
azure-core==1.31.0
azure-storage-blob==12.23.1
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==5.0.1
blinker==1.8.2
boto==2.49.0
boto3==1.35.49
botocore==1.35.49
boxing==0.1.4
Brlapi==0.8.3
cachetools==5.3.3
certifi==2020.6.20
cffi==1.17.0
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.5
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
dlib==19.24.6
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
duplicity==0.8.21
entrypoints==0.4
exceptiongroup==1.2.1
executing==2.0.1
expiringdict==1.2.2
face-recognition==1.3.0
face-recognition-models==0.3.0
fasteners==0.14.1
fastjsonschema==2.20.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
future==0.18.2
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.7.2
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==1.1.2
groq==0.13.1
grpc-google-iam-v1==0.13.1
grpcio==1.65.0
grpcio-status==1.62.2
gspread==6.1.4
gyp==0.1
h11==0.14.0
h5py==3.11.0
htmlmin==0.1.12
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
humansignal-drf-yasg==1.21.9
idna==3.3
ijson==3.3.0
imageio==2.35.0
importlib_metadata==8.4.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.26.0
ipython-genutils==0.2.0
ipywidgets==8.1.3
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonpointer==3.0.0
jsonschema==3.2.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.11
keras==3.5.0
keras-lr-finder==0.1
keras-vis==0.4.1
keyring==23.5.0
kiwisolver==1.4.5
label-studio==1.13.1
label-studio-sdk==1.0.5
labelImg==1.8.6
language-selector==0.1
launchdarkly-server-sdk==8.2.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==18.1.1
lockfile==0.12.2
louis==3.20.0
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.15
mistune==3.0.2
ml-dtypes==0.4.0
mlxtend==0.23.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.5
namex==0.0.8
narwhals==1.18.4
nbclassic==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
nltk==3.9.1
notebook==6.5.7
notebook_shim==0.2.4
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
openai==0.28.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opentelemetry-api==1.27.0
opt-einsum==3.3.0
optree==0.12.1
ordered-set==4.0.2
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
PyDrive==1.3.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pyRFC3339==1.1
pyrsistent==0.20.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-json-logger==2.0.4
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.2
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
redis==3.5.3
referencing==0.35.1
regex==2024.9.11
reportlab==3.6.8
requests==2.32.3
requests-mock==1.12.1
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
rq==1.10.1
rsa==4.9
rules==2.2
s3transfer==0.10.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
SecretStorage==3.3.1
semver==3.0.2
Send2Trash==1.8.3
sentry-sdk==2.17.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.5
SpeechRecognition==3.12.0
SQLAlchemy==1.4.31
sqlparse==0.5.1
stack-data==0.6.3
streamlit==1.41.1
systemd-python==234
tenacity==9.0.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminado==0.18.1
textblob==0.18.0.post0
threadpoolctl==3.5.0
tifffile==2024.8.10
tinycss2==1.3.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.20
usb-creator==0.3.7
user-agents==2.2.0
vis==0.0.5
visualkeras==0.1.3
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
xdg==5
xkit==0.0.0
xmljson==0.2.1
yarl==1.9.4
zipp==1.0.0
